Batch wall label refreshing in canvas pan callback
Wall updates were previously calculated each time the `canvasPan` event was triggered. This can happen multiple time in one frame and even more so multiple times between actual rerenders of the game canvas. Especially on high refresh rate displays with limited FPS in foundry, this can lead to many wasted calculations and dramatically impact performance when trying to pan the screen.

With this commit, wall identifier labels are not calculated immediately in the canvasPan event listener, but rather scheduled to be executed in the next game canvas render tick. Further invocations of the canvasPan callback while an update is scheduled are discarded.